requirements.txt
# requirements.txt (User Provided)
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kombu==5.5.2
lxml==5.3.2
Mako==1.3.9
MarkupSafe==3.0.2
openai==1.70.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stripe==12.0.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
XlsxWriter==3.2.2
zipp==3.21.0
